Fix check_mode for Tomcat Version Check
When running the Ansible in check mode, the task "Download & Extract
Tomcat" fails because the `tomcat_check` variable is not defined.

The cause is that when Ansible runs in check mode, tasks with the
'command' module are not executed. The Tomcat Version check task
executes a java command that registers the `tomcat_check` variable,
which is used in the next task ("Download & Extract Tomcat").

As the command is not executed, the variable does not exist and the
conditional test fails.

Forcing the execution of the command, even in check mode, makes the
variable to be defined and the checks to be successful.